Complete panorama GUI download
Start the download when button is clicked. Display download progress in a toplevel. Allow cancellation. Use threading to avoid freezing the GUI. Adapt API code to work with the GUI threading. Create the save image screen where an image preview is shown (with scrollbars if necessary) and the options of saving, returning to the main menu and returning back to download again. The save image screen can be used again for after a URL download, which is what must be implemented next (much simpler!).