Implement URL download
Allows the user to input a Google Street View URL (insta-validated) and width/height using both synchronised sliders and entries. There is a Download button which when pressed starts the download thread. Once the download completes the save image screen is displayed like before. The download can be cancelled, and will automatically be cancelled if the URL screen is exited or the app is closed (daemons).