Minor code cleanup
Reduce code repetition and fixed a few minor bugs relating to missing panorama data. Also check 32x32 top left area of a panorama in advance of cropping black edges (if all pixels are black in this region, assume the entire image is black so return an empty image). Performance can also be improved in terms of panorama stitching and cropping. C++?