Start main live downloading
Logger added, and Export Log feature implemented (basic - text file). When the start button is pressed, a Chrome window is launched and the maps website is loaded too. Basic stop system in place too already. Generally a stable interface between the selenium window and GUI, ready for the core logic.