Keybind download near-completion + logger enhancements
Add code to allow keybind to be registered and thus the URL captured (error if invalid, otherwise downloaded). The keybind must match the settings keybind, the keybind capture mode must be active and the tracked chrome window must be in the fg at the time of checking. Some degree of robustness and corner case of Ctrl + <key> and even Ctrl + Alt + <key> handled reasonably. Improve logger quality including line limit (delete old lines, avoids infinitely growing in case user uses the program for extended time) and also auto-scroll to the bottom when the scrollbar is at the bottom. V.good