General live downloader improvements
Upon setting capture mode to 'Fixed Intervals', the time until next capture resets to the time interval, rather than continuing from the previous time since this makes more sense. Remove 'Chrome is being automated' message (unnecessary). Consider keybind invalid URL as an error which may stop live tracking if 'Stop upon error' is on. The default keybind has been changed from 's' to 'q' since 'q' does not appear to control anything in Google Maps and is in a convenient position on keyboard. When downlode mode is set to 'On Demand', the queue is shrinked to size 1 (most recently added) to adjust to the setting change. This occurs when the download mode is set to 'Queue', URLs accumulate in the queue, but when the setting is changed to 'on demand', this should no longer be the case. Also a way to empty the queue except the most recently added URL. Not much code added, but good robustness and Quality of Life improvements nonetheless.