Minor Refactors 2 - Done
Main changes include moving all C++ to the GUI folder (more sense). Also fix minor bugs in the Batch processing table and make the scrollbars only appear when needed. Other minor changes to code and some comments scattered throughout the harder to understand parts. Tested all components as part of terminal testing. Project virtually finished for now! Just need to package and document and that should be the end of the project.